refactor(Bot): 重构语音消息处理逻辑
- 移除 api.py 中的直接处理逻辑，改为调用 save_bbox 函数- 注释掉 api.py 中的多余代码，简化处理流程
- 在 voice_forward.py 中添加循环检测 data.txt 文件存在的逻辑
-优化 voice_forward.py 中的文件读取和语音发送流程

+++ Bot/OtherUse/voice_forward.py
@@ -22,12 +22,19 @@ def send_voice(msg):
 
 if __name__ == '__main__':
     # 定时任务每两小时执行一次
-    time.sleep(7200)
+    while True:
+        time.sleep(10)
+        # 如果data.txt存在，则发送
+        if os.path.exists('data.txt'):
+            print('data.txt存在')
+            with open('data.txt', 'r') as f:
+                lines = f.readlines()
+            for line in lines:
+                send_voice(line)
 
-    with open('data.txt', 'r') as f:
-        lines = f.readlines()
+            os.remove('data.txt')
+        else:
+            print('data.txt不存在')
 
-    for line in lines:
-        send_voice(line)
 
-    os.remove('data.txt')
+
